TJDMX: - Splitted off the actual DMX engine into a separate library (TJDMXEngine); TJDMXEngine is more or less portable (i.e. only depends on TJSharedUI on Windows; some devices are disabled when building on other platforms, but ArtNet, ESP and LanBox work).
TJNP:
- Added SendTo and more variants of Send in Socket class
- Added Socket::SetBroadcast/SetReuseAddress/SetBlocking
- WebServer now uses Bytes to identify item sizes

TJShared:
- Data now uses Bytes to indicate size (which is a 64 bit integer)

git-svn-id: https://svn.pixelspark.nl/svn/corespark/trunk@664 c9a305e2-9f8a-4be3-8c95-b0e5bab41dc9

+++ TJNP/src/tjhttp.cpp
@@ -170,7 +170,7 @@ HTTPRequest::HTTPRequest(ref<Data> request, ref<Data> data) {
 
 		// Find where the headers end
 		const char* fullRequest = request->GetBuffer();
-		const std::string req(fullRequest, 0, request->GetSize());
+		const std::string req(fullRequest, 0, (size_t)request->GetSize());
 		std::string currentHeader;
 		std::string::const_iterator it = req.begin();
